+++ drivers/duckdb/duckdb.go
@@ -569,14 +569,14 @@ func (d *driveri) CreateTable(ctx context.Context, db sqlz.DB, tblDef *schema.Ta
 
 // CreateSchema implements driver.SQLDriver.
 func (d *driveri) CreateSchema(ctx context.Context, db sqlz.DB, schemaName string) error {
-	stmt := fmt.Sprintf(`CREATE SCHEMA %q`, schemaName)
+	stmt := "CREATE SCHEMA " + stringz.DoubleQuote(schemaName)
 	_, err := db.ExecContext(ctx, stmt)
 	return errz.Wrapf(errw(err), "duckdb: create schema {%s}", schemaName)
 }
 
 // DropSchema implements driver.SQLDriver.
 func (d *driveri) DropSchema(ctx context.Context, db sqlz.DB, schemaName string) error {
-	stmt := fmt.Sprintf(`DROP SCHEMA %q CASCADE`, schemaName)
+	stmt := "DROP SCHEMA " + stringz.DoubleQuote(schemaName) + " CASCADE"
 	_, err := db.ExecContext(ctx, stmt)
 	return errz.Wrapf(errw(err), "duckdb: drop schema {%s}", schemaName)
 }
@@ -762,6 +762,11 @@ func (g *grip) Source() *source.Source {
 // SourceMetadata implements driver.Grip.
 func (g *grip) SourceMetadata(ctx context.Context, noSchema bool) (*metadata.Source, error) {
 	return getSourceMetadata(ctx, g.src, g.db, noSchema)
+}
+
+// DBSemver implements driver.Grip.
+func (g *grip) DBSemver(ctx context.Context) (string, error) {
+	return g.drvr.DBSemver(ctx, g.db)
 }
 
 // TableMetadata implements driver.Grip.